+++ src/main/java/ru/yandex/practicum/filmorate/model/Feed.java
@@ -5,11 +5,11 @@ import lombok.*;
 @Data
 public class Feed {
     private final Long timestamp;
-    private final int userId;
+    private final long userId;
     private final EventType eventType;
     private final Operation operation;
-    private final int eventId;
-    private final int entityId;
+    private final long eventId;
+    private final long entityId;
 
     public enum EventType {
         LIKE,
@@ -23,3 +23,4 @@ public class Feed {
         UPDATE
     }
 }
+
